+++ terraform/infra/monitoring/dashboard.tf
@@ -1,4 +1,6 @@
 resource "helm_release" "kubernetes_dashboard" {
+  count = var.hostname == "kubernetes" ? 1 : 0 #only install for dev stages
+
   name       = "kubernetes-dashboard"
   repository = "https://kubernetes.github.io/dashboard"
   chart      = "kubernetes-dashboard"
@@ -26,6 +28,8 @@ resource "helm_release" "kubernetes_dashboard" {
 }
 
 resource "kubernetes_manifest" "dashboard-role" {
+  count = var.hostname == "kubernetes" ? 1 : 0
+
   manifest   = yamldecode(
     <<-EOF
     apiVersion: rbac.authorization.k8s.io/v1
@@ -45,6 +49,8 @@ resource "kubernetes_manifest" "dashboard-role" {
 }
 
 resource "kubernetes_manifest" "dashboard-ingress" {
+  count = var.hostname == "kubernetes" ? 1 : 0
+
   manifest   = yamldecode(
   <<-EOF
   kind: Ingress
@@ -54,8 +60,6 @@ resource "kubernetes_manifest" "dashboard-ingress" {
     namespace: monitoring
     annotations:
       cert-manager.io/cluster-issuer: my-cluster-issuer
-      nginx.ingress.kubernetes.io/auth-secret: authentication-secret
-      nginx.ingress.kubernetes.io/auth-type: basic
       nginx.ingress.kubernetes.io/backend-protocol: HTTPS
       nginx.ingress.kubernetes.io/rewrite-target: /$1
       service.alpha.kubernetes.io/app-protocols: '{"https":"HTTPS"}'